Accessibility Update
Elements that behave as buttons but are built using other tags such as span, div, a or others, now include a "role" attribute that equals to "button". Elements that visually appear as titles but are coded with a non-heading HTML Tag now include a "role" attribute that equals to "heading" or have their tags fixed. Menu items that have a dropdown menu now include an "aria-haspopup" attribute that equals to "true".