Drop Down Menu Fix
moved role=button label from a to li for the dropdown list items, this is because of the need for a step-down functionality which clicks the a tag inside the li items when pressing enter. Meaning that while the li element is not actually a button, screen readers will now view them as such to avoid confusion and aid accessibility.